Adjusted the models structure and added last_login update function to user model
