+++ newspaper/forms.py
@@ -1,6 +1,6 @@
 from django.contrib.auth.forms import UserCreationForm
 from django import forms
-from .models import User
+from newspaper.models import User
 
 class ReaderCreationForm(UserCreationForm):
         class Meta:
