+++ newspaper/models/__init__.py
@@ -1,0 +1,2 @@
+from .newspaper_models import *
+from .user_models import *
