+++ newspaper/views/article_views.py
@@ -20,7 +20,7 @@ from newspaper.serializers import ArticleSerializer, UserSerializer
 # Article views
 
 @api_view(['GET'])
-@permission_classes([IsAuthenticated])
+# @permission_classes([IsAuthenticated])
 def articlesView(request):
     articles = models.Article.objects.all()
     articles_serializer = ArticleSerializer(articles, many=True)
