+++ newspaper_project/settings.py
@@ -1,23 +1,9 @@
-"""
-Django settings for newspaper_project project.
-
-Generated by 'django-admin startproject' using Django 4.2.3.
-
-For more information on this file, see
-https://docs.djangoproject.com/en/4.2/topics/settings/
-
-For the full list of settings and their values, see
-https://docs.djangoproject.com/en/4.2/ref/settings/
-"""
-
 from pathlib import Path
 from dotenv import load_dotenv
 import os
 from datetime import timedelta
 
 load_dotenv()
-
-
 
 
 # Build paths inside the project like this: BASE_DIR / 'subdir'.
