+++ newspaper_project/urls.py
@@ -19,8 +19,10 @@ from django.urls import path, include
 from rest_framework_simplejwt.views import TokenObtainPairView, TokenRefreshView
 
 urlpatterns = [
-    path('admin/', admin.site.urls),
     path('', include('newspaper.urls')),
-    # path('api/token/', TokenObtainPairView.as_view(), name='token_obtain_pair'),
+    path('admin/', admin.site.urls),
     path('token/refresh/', TokenRefreshView.as_view(), name='token_refresh')
+    
+    # included in user logging in function view
+    # path('api/token/', TokenObtainPairView.as_view(), name='token_obtain_pair'),
 ]
